correction covariance estimation
- During generating ROM_PIV files [B0001_new.dat], ascii files [U*.txt]
are stored in a [cov_after_gaussSmoothing] directory for future
covariance estimation.

These [U*.txt] files contain velocity components for each point of the
smooth/crop plane Z=cte; associated coordinates are stored in
[XYcrop.txt] file, and list of this latter file and other [U*.txt] files
are written in the [list_fic_time.txt] file

- After generating ROM_PIV files [B0001_new.dat] for each time,
covariance is estimated via cov_after_gaussSmoothing_stat.C code, before
defining InvCov.

NOTE : with a classic slice paraFoam operation, time evolution is
available. But, it's not the case when this slice is applied on a
smooth/crop box; and so, temporal statistic can't be applied !